requirements.txt
APScheduler==3.6.3
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
paramiko==2.7.2
pkg-resources==0.0.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.15.0
python-telegram-bot==13.1
pytz==2020.5
six==1.15.0
tornado==6.1
tzlocal==2.1
